+++ R/OptimInstance.R
@@ -107,7 +107,9 @@ OptimInstance = R6Class("OptimInstance",
     #'   columns for extra information.
     #' @param y (`numeric(1)`)\cr
     #'   Optimal outcome.
-    assign_result = function(xdt, y) {
+    #' @param ... (`any`)\cr
+    #' ignored.
+    assign_result = function(xdt, y, ...) {
       stop("Abstract class")
     },
 
@@ -152,7 +154,11 @@ OptimInstance = R6Class("OptimInstance",
   ),
 
   private = list(
+    # intermediate objects
+    .result_xdt = NULL,
+    .result_extra = NULL,
     .result = NULL,
+
     .label = NULL,
     .man = NULL,