+++ R/zzz.R
@@ -5,6 +5,7 @@
 #' @importFrom R6 R6Class
 #' @importFrom utils capture.output head tail
 #' @importFrom methods formalArgs
+#' @importFrom stats setNames
 "_PACKAGE"
 
 .onLoad = function(libname, pkgname) {
@@ -33,5 +34,7 @@
     lg$set_threshold("warn")
   }
 } # nocov end
+
+utils::globalVariables("batch_nr")
 
 leanify_package()