+++ tests/testthat/test_OptimInstanceAsyncSingleCrit.R
@@ -1,6 +1,5 @@
 test_that("initializing OptimInstanceAsyncSingleCrit works", {
   skip_on_cran()
-  #  skip_on_cran()
   skip_if_not_installed("rush")
   flush_redis()
 
@@ -24,11 +23,10 @@ test_that("initializing OptimInstanceAsyncSingleCrit works", {
 
 test_that("rush controller can be passed to OptimInstanceAsyncSingleCrit", {
   skip_on_cran()
-  #  skip_on_cran()
   skip_if_not_installed("rush")
   flush_redis()
 
-  rush = rsh(network_id = "remote_network")
+  rush = rush::rsh(network_id = "remote_network")
 
   instance = oi_async(
     objective = OBJ_2D,
@@ -43,11 +41,50 @@ test_that("rush controller can be passed to OptimInstanceAsyncSingleCrit", {
 
 test_that("context is initialized correctly", {
   skip_on_cran()
-  #  skip_on_cran()
   skip_if_not_installed("rush")
   flush_redis()
 
   rush::rush_plan(n_workers = 2)
+  instance = oi_async(
+    objective = OBJ_2D,
+    search_space = PS_2D,
+    terminator = trm("evals", n_evals = 5L)
+  )
+
+  optimizer = opt("async_random_search")
+  optimizer$optimize(instance)
+
+  expect_r6(instance$objective$context, "ContextAsync")
+})
+
+test_that("point evaluation works", {
+  skip_on_cran()
+  skip_if_not_installed("rush")
+  flush_redis()
+
+  # evaluation takes place on the workers
+  rush = rush::RushWorker$new("test", remote = FALSE)
+  instance = oi_async(
+    objective = OBJ_2D,
+    search_space = PS_2D,
+    terminator = trm("evals", n_evals = 5L),
+    rush = rush
+  )
+
+  expect_equal(get_private(instance)$.eval_point(list(x1 = 1, x2 = 0)), list(y = 1))
+})
+
+test_that("reconnect method works", {
+  skip_on_cran()
+  skip_if_not_installed("rush")
+  flush_redis()
+
+  on.exit({
+    file.remove("instance.rds")
+  })
+
+  rush::rush_plan(n_workers = 2)
+
   instance = oi_async(
     objective = OBJ_2D,
     search_space = PS_2D,
@@ -57,5 +94,10 @@ test_that("context is initialized correctly", {
   optimizer = opt("async_random_search")
   optimizer$optimize(instance)
 
-  expect_r6(instance$objective$context, "ContextAsync")
+  saveRDS(instance, file = "instance.rds")
+  instance = readRDS("instance.rds")
+
+  instance$reconnect()
+
+  expect_r6(instance, "OptimInstanceAsyncSingleCrit")
 })