+++ tests/testthat/test_OptimizerAsyncRandomSearch.R
@@ -1,6 +1,5 @@
 test_that("OptimizerAsyncRandomSearch works", {
   skip_on_cran()
-  #  skip_on_cran()
   skip_if_not_installed("rush")
   flush_redis()